Added more versatility to the LUT generator. Fixed the LUT so that time-bins are correct.

+++ 2017/LUT.h
@@ -19,7 +19,7 @@ typedef struct lut * LUT;
 
 struct lut {
     int nrows;  ///< Number of rows.
-    unsigned int *times;   ///< Array of times.
+    double *times;   ///< Array of times.
     double *vals;    ///< Array of values.
 };
